Work on basler camera timestamps
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@26006 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 1b98170b5d3d6abeb8ed1ebc67e15a7fd9d44e8a

+++ wrapper/baslercamera/src/de/uniol/inf/is/odysseus/wrapper/baslercamera/swig/BaslerCamera.java
@@ -83,11 +83,11 @@ public class BaslerCamera {
     return BaslerJavaJNI.BaslerCamera_getImageHeight(swigCPtr, this);
   }
 
-  public long getLastTimeStamp() {
+  public double getLastTimeStamp() {
     return BaslerJavaJNI.BaslerCamera_getLastTimeStamp(swigCPtr, this);
   }
 
-  public void onGrabbed(long timeStamp, java.nio.ByteBuffer buffer) {
+  public void onGrabbed(double timeStamp, java.nio.ByteBuffer buffer) {
     if (getClass() == BaslerCamera.class) BaslerJavaJNI.BaslerCamera_onGrabbed(swigCPtr, this, timeStamp, buffer); else BaslerJavaJNI.BaslerCamera_onGrabbedSwigExplicitBaslerCamera(swigCPtr, this, timeStamp, buffer);
   }
